+++ ts/screens/wallet/TransactionDetailsScreenOld.tsx
@@ -26,7 +26,7 @@ import {
 import IconFont from "../../components/ui/IconFont";
 import Markdown from "../../components/ui/Markdown";
 import { RotatedCards } from "../../components/wallet/card/RotatedCards";
-import WalletLayoutNew from "../../components/wallet/WalletLayoutNew";
+import WalletLayout from "../../components/wallet/WalletLayout";
 import I18n from "../../i18n";
 import { navigateToWalletHome } from "../../store/actions/navigation";
 import { Dispatch } from "../../store/actions/types";
@@ -90,21 +90,13 @@ const styles = StyleSheet.create({
   }
 });
 
-class TransactionDetailsScreenNew extends React.Component<Props> {
-  private displayedWallet(transactionWallet: any) {
-    return transactionWallet ? (
-      <RotatedCards cardType="Preview" wallets={[transactionWallet]} />
-    ) : (
-      <RotatedCards cardType="Preview" />
-    );
-  }
-
+class TransactionDetailsScreenOld extends React.Component<Props> {
   /**
    * It provides the proper header to the screen. If isTransactionStarted
    * (the user displays the screen during the process of identify and accept a transaction)
    * then the "Thank you message" is displayed
    */
-  private topContent(paymentCompleted: boolean, transactionWallet: any) {
+  private getSubHeader(paymentCompleted: boolean) {
     return paymentCompleted ? (
       <View>
         <Grid>
@@ -121,7 +113,6 @@ class TransactionDetailsScreenNew extends React.Component<Props> {
           </Col>
           <Col size={1} />
         </Grid>
-        {this.displayedWallet(transactionWallet)}
       </View>
     ) : (
       <View spacer={true} />
@@ -169,19 +160,24 @@ class TransactionDetailsScreenNew extends React.Component<Props> {
       centsToAmount(transaction.grandTotal.amount)
     );
 
+    // FIXME: in case the wallet for this transaction has been deleted, display
+    //        a message in the wallet layout instead of an empty space
     const transactionWallet = this.props.wallets
       ? this.props.wallets[transaction.idWallet]
       : undefined;
 
     return (
-      <WalletLayoutNew
+      <WalletLayout
         title={I18n.t("wallet.transaction")}
+        headerContents={this.getSubHeader(isPaymentCompletedTransaction)}
+        displayedWallets={
+          transactionWallet ? (
+            <RotatedCards cardType="Preview" wallets={[transactionWallet]} />
+          ) : (
+            <RotatedCards cardType="Preview" />
+          )
+        }
         allowGoBack={!isPaymentCompletedTransaction}
-        topContent={this.topContent(
-          isPaymentCompletedTransaction,
-          transactionWallet
-        )}
-        hideHeader={true}
       >
         <Content
           scrollEnabled={false}
@@ -234,7 +230,7 @@ class TransactionDetailsScreenNew extends React.Component<Props> {
             )}
           </Grid>
         </Content>
-      </WalletLayoutNew>
+      </WalletLayout>
     );
   }
 }
@@ -252,7 +248,7 @@ export default connect(
   mapDispatchToProps
 )(
   withContextualHelp(
-    TransactionDetailsScreenNew,
+    TransactionDetailsScreenOld,
     I18n.t("wallet.whyAFee.title"),
     () => <Markdown>{I18n.t("wallet.whyAFee.text")}</Markdown>
   )
