+++ ts/screens/wallet/TransactionsScreenOld.tsx
@@ -11,7 +11,7 @@ import { connect } from "react-redux";
 
 import CardComponent from "../../components/wallet/card/CardComponent";
 import TransactionsList from "../../components/wallet/TransactionsList";
-import WalletLayoutNew from "../../components/wallet/WalletLayoutNew";
+import WalletLayout from "../../components/wallet/WalletLayout";
 import I18n from "../../i18n";
 import {
   navigateToTransactionDetailsScreen,
@@ -69,51 +69,42 @@ const ListEmptyComponent = (
   </Content>
 );
 
-class TransactionsScreenNew extends React.Component<Props> {
-  private headerContent(
-    selectedWallet: Wallet,
-    isFavorite: pot.Pot<boolean, Error>
-  ) {
-    return (
-      <React.Fragment>
-        <View>
-          <View style={styles.walletBannerText}>
-            <Text white={true}>{I18n.t("wallet.creditDebitCards")}</Text>
-          </View>
-          <View spacer={true} />
-        </View>
-
-        <CardComponent
-          type="Header"
-          wallet={selectedWallet}
-          hideFavoriteIcon={false}
-          hideMenu={false}
-          isFavorite={isFavorite}
-          onSetFavorite={(willBeFavorite: boolean) =>
-            this.props.setFavoriteWallet(
-              willBeFavorite ? selectedWallet.idWallet : undefined
-            )
-          }
-          onDelete={() => this.props.deleteWallet(selectedWallet.idWallet)}
-        />
-      </React.Fragment>
-    );
-  }
-
+class TransactionsScreenOld extends React.Component<Props> {
   public render(): React.ReactNode {
     const selectedWallet = this.props.navigation.getParam("selectedWallet");
-
     const isFavorite = pot.map(
       this.props.favoriteWallet,
       _ => _ === selectedWallet.idWallet
     );
+    const headerContents = (
+      <View>
+        <View style={styles.walletBannerText}>
+          <Text white={true}>{I18n.t("wallet.creditDebitCards")}</Text>
+        </View>
+        <View spacer={true} />
+      </View>
+    );
 
     return (
-      <WalletLayoutNew
+      <WalletLayout
         title={I18n.t("wallet.paymentMethod")}
         allowGoBack={true}
-        topContent={this.headerContent(selectedWallet, isFavorite)}
-        hideHeader={true}
+        headerContents={headerContents}
+        displayedWallets={
+          <CardComponent
+            type="Header"
+            wallet={selectedWallet}
+            hideFavoriteIcon={false}
+            hideMenu={false}
+            isFavorite={isFavorite}
+            onSetFavorite={(willBeFavorite: boolean) =>
+              this.props.setFavoriteWallet(
+                willBeFavorite ? selectedWallet.idWallet : undefined
+              )
+            }
+            onDelete={() => this.props.deleteWallet(selectedWallet.idWallet)}
+          />
+        }
       >
         <TransactionsList
           title={I18n.t("wallet.transactions")}
@@ -124,7 +115,7 @@ class TransactionsScreenNew extends React.Component<Props> {
           }
           ListEmptyComponent={ListEmptyComponent}
         />
-      </WalletLayoutNew>
+      </WalletLayout>
     );
   }
 }
@@ -168,4 +159,4 @@ const mapDispatchToProps = (dispatch: Dispatch) => ({
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(TransactionsScreenNew);
+)(TransactionsScreenOld);
