+++ ts/screens/wallet/WalletHomeScreen.tsx
@@ -3,22 +3,22 @@
  * a "pay notice" button and payment methods info/button to
  * add new ones
  */
+import { none } from "fp-ts/lib/Option";
 import * as pot from "italia-ts-commons/lib/pot";
-import { Button, Content, H1, Left, Right, Text, View } from "native-base";
+import { Button, Content, Text, View } from "native-base";
 import * as React from "react";
 import { Image, StyleSheet } from "react-native";
 import { Grid, Row } from "react-native-easy-grid";
 import { NavigationScreenProp, NavigationState } from "react-navigation";
 import { connect } from "react-redux";
 
-import { none } from "fp-ts/lib/Option";
 import BoxedRefreshIndicator from "../../components/ui/BoxedRefreshIndicator";
 import H5 from "../../components/ui/H5";
+import IconFont from "../../components/ui/IconFont";
 import { AddPaymentMethodButton } from "../../components/wallet/AddPaymentMethodButton";
 import CardsFan from "../../components/wallet/card/CardsFan";
 import TransactionsList from "../../components/wallet/TransactionsList";
 import WalletLayout from "../../components/wallet/WalletLayout";
-import { DEFAULT_APPLICATION_NAME } from "../../config";
 import I18n from "../../i18n";
 import {
   navigateToPaymentScanQrCode,
@@ -29,11 +29,12 @@ import {
 import { Dispatch } from "../../store/actions/types";
 import { fetchTransactionsRequest } from "../../store/actions/wallet/transactions";
 import { fetchWalletsRequest } from "../../store/actions/wallet/wallets";
+import { isPagoPATestEnabledSelector } from "../../store/reducers/persistedPreferences";
 import { GlobalState } from "../../store/reducers/types";
 import { latestTransactionsSelector } from "../../store/reducers/wallet/transactions";
 import { walletsSelector } from "../../store/reducers/wallet/wallets";
 import variables from "../../theme/variables";
-import { Transaction } from "../../types/pagopa";
+import { Transaction, Wallet } from "../../types/pagopa";
 
 type OwnProps = Readonly<{
   navigation: NavigationScreenProp<NavigationState>;
@@ -44,11 +45,6 @@ type Props = ReturnType<typeof mapStateToProps> &
   OwnProps;
 
 const styles = StyleSheet.create({
-  flex: {
-    alignItems: "flex-end",
-    justifyContent: "space-between"
-  },
-
   inLineSpace: {
     lineHeight: 20
   },
@@ -63,17 +59,25 @@ const styles = StyleSheet.create({
     justifyContent: "center",
     backgroundColor: "transparent"
   },
+
+  flex1: {
+    flex: 1
+  },
+
+  flexRow: {
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "space-between"
+  },
+
   emptyListWrapper: {
     padding: variables.contentPadding,
     alignItems: "center"
   },
+
   emptyListContentTitle: {
     paddingBottom: variables.contentPadding / 2,
     fontSize: variables.fontSizeSmall
-  },
-
-  brandDarkGray: {
-    color: variables.brandDarkGray
   },
 
   bordercColorBrandGray: {
@@ -84,9 +88,16 @@ const styles = StyleSheet.create({
     color: variables.brandGray
   },
 
-  whiteContent: {
-    backgroundColor: variables.colorWhite,
-    flex: 1
+  brandDarkGray: {
+    color: variables.brandDarkGray
+  },
+
+  brandLightGray: {
+    color: variables.brandLightGray
+  },
+
+  whiteBg: {
+    backgroundColor: variables.colorWhite
   },
 
   noBottomPadding: {
@@ -102,58 +113,53 @@ const styles = StyleSheet.create({
   }
 });
 
-const ListEmptyComponent = (
-  <Content scrollEnabled={false} noPadded={true}>
-    <View style={styles.emptyListWrapper}>
-      <Text style={styles.emptyListContentTitle}>
-        {I18n.t("wallet.noTransactionsInWalletHome")}
-      </Text>
-      <Image
-        source={require("../../../img/messages/empty-transaction-list-icon.png")}
-      />
-    </View>
-  </Content>
-);
-
 /**
  * Wallet Home Screen
  */
 class WalletHomeScreen extends React.Component<Props, never> {
-  private header() {
+  public componentDidMount() {
+    // WIP loadTransactions should not be called from here
+    // (transactions should be persisted & fetched periodically)
+    // WIP WIP create pivotal story
+    this.props.loadWallets();
+    this.props.loadTransactions();
+  }
+
+  private cardHeader() {
     return (
-      <Row style={styles.flex}>
-        <H1 style={styles.white}>{I18n.t("wallet.wallet")}</H1>
-        <Image source={require("../../../img/wallet/bank.png")} />
-      </Row>
+      <View style={styles.flexRow}>
+        <View>
+          <H5 style={styles.brandLightGray}>
+            {I18n.t("wallet.paymentMethods")}
+          </H5>
+        </View>
+        <View>
+          <AddPaymentMethodButton
+            onPress={this.props.navigateToWalletAddPaymentMethod}
+          />
+        </View>
+      </View>
     );
   }
 
-  private withCardsHeader() {
+  private cardPreview(wallets: any) {
     return (
-      <Grid>
-        {this.header()}
+      <View>
+        {this.cardHeader()}
         <View spacer={true} />
-        <Row>
-          <Left>
-            <Text bold={true} white={true}>
-              {I18n.t("wallet.paymentMethods")}
-            </Text>
-          </Left>
-          <Right>
-            <AddPaymentMethodButton
-              onPress={this.props.navigateToWalletAddPaymentMethod}
-            />
-          </Right>
-        </Row>
-      </Grid>
+        <CardsFan
+          wallets={
+            wallets.length === 1 ? [wallets[0]] : [wallets[0], wallets[1]]
+          }
+          navigateToWalletList={this.props.navigateToWalletList}
+        />
+      </View>
     );
   }
 
   private withoutCardsHeader() {
     return (
       <Grid>
-        {this.header()}
-        <View spacer={true} />
         <Row>
           <Text note={true} white={true} style={styles.inLineSpace}>
             {I18n.t("wallet.newPaymentMethod.addDescription")}
@@ -186,8 +192,6 @@ class WalletHomeScreen extends React.Component<Props, never> {
   private loadingWalletsHeader() {
     return (
       <View>
-        {this.header()}
-        <View spacer={true} />
         <BoxedRefreshIndicator
           caption={
             <Text white={true} style={styles.inLineSpace}>
@@ -199,39 +203,11 @@ class WalletHomeScreen extends React.Component<Props, never> {
     );
   }
 
-  private fixedSubHeader() {
-    return (
-      <View>
-        <View spacer={true} />
-        <View style={styles.animatedSubHeaderContent}>
-          <H5 style={styles.brandDarkGray}>
-            {I18n.t("wallet.latestTransactions")}
-          </H5>
-          <Text>{I18n.t("wallet.total")}</Text>
-        </View>
-        <View spacer={true} />
-      </View>
-    );
-  }
-
-  // TODO: insert a more detailed value for appheader height
-  private headerHeight: number = 42 + 90;
-  private AnimatedViewHeight: number =
-    this.headerHeight + (variables.h5LineHeight + 2 * variables.spacerWidth); //
-  // TODO: insert a more detailed value for topContentHeight
-  private topContentHeight: number = 450;
-  private topContentHeightOffset: number = 40;
-  private interpolationVars: ReadonlyArray<number> = [
-    this.AnimatedViewHeight,
-    this.topContentHeight,
-    this.topContentHeightOffset
-  ];
-
   private errorWalletsHeader() {
     return (
       <View>
-        {this.withCardsHeader()}
-        <View spacer={true} large={true} />
+        {this.cardHeader()}
+        <View spacer={true} />
         <Text style={[styles.white, styles.inLineSpace]}>
           {I18n.t("wallet.walletLoadFailure")}
         </Text>
@@ -250,84 +226,137 @@ class WalletHomeScreen extends React.Component<Props, never> {
     );
   }
 
-  public componentDidMount() {
-    // WIP loadTransactions should not be called from here
-    // (transactions should be persisted & fetched periodically)
-    // WIP WIP create pivotal story
-    this.props.loadWallets();
-    this.props.loadTransactions();
+  private fixedSubHeader() {
+    return (
+      <View style={styles.whiteBg}>
+        <View spacer={true} />
+        <View style={styles.animatedSubHeaderContent}>
+          <H5 style={styles.brandDarkGray}>
+            {I18n.t("wallet.latestTransactions")}
+          </H5>
+          <Text>{I18n.t("wallet.total")}</Text>
+        </View>
+        <View spacer={true} />
+      </View>
+    );
   }
 
-  public render(): React.ReactNode {
-    const { potWallets, potTransactions } = this.props;
-    const wallets = pot.getOrElse(potWallets, []);
-    const headerContents = pot.isLoading(potWallets)
-      ? this.loadingWalletsHeader()
-      : pot.isError(potWallets)
-        ? this.errorWalletsHeader()
-        : wallets.length > 0
-          ? this.withCardsHeader()
-          : this.withoutCardsHeader();
-
+  private transactionError() {
     return (
-      <WalletLayout
-        title={DEFAULT_APPLICATION_NAME}
-        headerContents={headerContents}
-        displayedWallets={
-          wallets.length === 0 ? null : (
-            <CardsFan
-              wallets={
-                wallets.length === 1 ? [wallets[0]] : [wallets[0], wallets[1]]
-              }
-              navigateToWalletList={this.props.navigateToWalletList}
-            />
-          )
+      <Content
+        scrollEnabled={false}
+        style={[styles.noBottomPadding, styles.whiteBg, styles.flex1]}
+      >
+        <View spacer={true} />
+        <H5 style={styles.brandDarkGray}>{I18n.t("wallet.transactions")}</H5>
+        <View spacer={true} large={true} />
+        <Text style={[styles.inLineSpace, styles.brandDarkGray]}>
+          {I18n.t("wallet.transactionsLoadFailure")}
+        </Text>
+        <View spacer={true} />
+        <Button
+          block={true}
+          light={true}
+          bordered={true}
+          small={true}
+          onPress={this.props.loadTransactions}
+        >
+          <Text primary={true}>{I18n.t("global.buttons.retry")}</Text>
+        </Button>
+        <View spacer={true} large={true} />
+      </Content>
+    );
+  }
+
+  private listEmptyComponent() {
+    return (
+      <Content scrollEnabled={false} noPadded={true}>
+        <View style={styles.emptyListWrapper}>
+          <Text style={styles.emptyListContentTitle}>
+            {I18n.t("wallet.noTransactionsInWalletHome")}
+          </Text>
+          <Image
+            source={require("../../../img/messages/empty-transaction-list-icon.png")}
+          />
+        </View>
+      </Content>
+    );
+  }
+
+  private transactionList(
+    potTransactions: pot.Pot<ReadonlyArray<Transaction>, Error>
+  ) {
+    return (
+      <TransactionsList
+        title={I18n.t("wallet.latestTransactions")}
+        totalAmount={I18n.t("wallet.total")}
+        transactions={potTransactions}
+        navigateToTransactionDetails={
+          this.props.navigateToTransactionDetailsScreen
         }
-        onNewPaymentPress={
+        ListEmptyComponent={this.listEmptyComponent()}
+      />
+    );
+  }
+
+  private footerButton(potWallets: pot.Pot<ReadonlyArray<Wallet>, Error>) {
+    return (
+      <Button
+        block={true}
+        onPress={
           pot.isSome(potWallets)
             ? this.props.navigateToPaymentScanQrCode
             : undefined
         }
+      >
+        <IconFont name="io-qr" style={styles.white} />
+        <Text>{I18n.t("wallet.payNotice")}</Text>
+      </Button>
+    );
+  }
+
+  /**
+   *  TODO: insert a more detailed value for appheader height 
+   * (see blocker: https://www.pivotaltracker.com/n/projects/2048617/stories/166154544)
+   */
+  private animatedViewHeight: number =
+    variables.h5LineHeight + 2 * variables.spacerWidth; //
+  // TODO: insert a more detailed value for topContentHeight
+  private topContentHeight: number = 450;
+  private topContentHeightOffset: number = 40;
+  private interpolationVars: ReadonlyArray<number> = [
+    this.animatedViewHeight,
+    this.topContentHeight,
+    this.topContentHeightOffset
+  ];
+
+  public render(): React.ReactNode {
+    const { potWallets, potTransactions } = this.props;
+    const wallets = pot.getOrElse(potWallets, []);
+    const headerContent = pot.isLoading(potWallets)
+      ? this.loadingWalletsHeader()
+      : pot.isError(potWallets)
+        ? this.errorWalletsHeader()
+        : wallets.length > 0
+          ? this.cardPreview(wallets)
+          : this.withoutCardsHeader();
+
+    const transactionContent = pot.isError(potTransactions)
+      ? this.transactionError()
+      : this.transactionList(potTransactions);
+
+    const footerContent = this.footerButton(potWallets);
+
+    return (
+      <WalletLayout
+        title={I18n.t("wallet.wallet")}
         allowGoBack={false}
         fixedSubHeader={this.fixedSubHeader()}
         interpolationVars={this.interpolationVars}
+        topContent={headerContent}
+        footerContent={footerContent}
       >
-        {pot.isError(potTransactions) ? (
-          <Content
-            scrollEnabled={false}
-            style={[styles.noBottomPadding, styles.whiteContent]}
-          >
-            <View spacer={true} />
-            <H5 style={styles.brandDarkGray}>
-              {I18n.t("wallet.transactions")}
-            </H5>
-            <View spacer={true} large={true} />
-            <Text style={[styles.inLineSpace, styles.brandDarkGray]}>
-              {I18n.t("wallet.transactionsLoadFailure")}
-            </Text>
-            <View spacer={true} />
-            <Button
-              block={true}
-              light={true}
-              bordered={true}
-              small={true}
-              onPress={this.props.loadTransactions}
-            >
-              <Text primary={true}>{I18n.t("global.buttons.retry")}</Text>
-            </Button>
-            <View spacer={true} large={true} />
-          </Content>
-        ) : (
-          <TransactionsList
-            title={I18n.t("wallet.latestTransactions")}
-            totalAmount={I18n.t("wallet.total")}
-            transactions={potTransactions}
-            navigateToTransactionDetails={
-              this.props.navigateToTransactionDetailsScreen
-            }
-            ListEmptyComponent={ListEmptyComponent}
-          />
-        )}
+        {transactionContent}
       </WalletLayout>
     );
   }
@@ -335,7 +364,8 @@ class WalletHomeScreen extends React.Component<Props, never> {
 
 const mapStateToProps = (state: GlobalState) => ({
   potWallets: walletsSelector(state),
-  potTransactions: latestTransactionsSelector(state)
+  potTransactions: latestTransactionsSelector(state),
+  isPagoPATestEnabled: isPagoPATestEnabledSelector(state)
 });
 
 const mapDispatchToProps = (dispatch: Dispatch) => ({
